Change the way that equality is implemented
